real-estate-app-with-web-scrapping-and-streamlit/requirements.txt
streamlit==1.31.1
pillow==10.2.0
plotly==5.19.0
pandas==2.2.1
beautifulsoup4==4.12.3